requirements.txt
streamlit
google-genai
groq
Pillow
requests
PyPDF2
pandas
beautifulsoup4
streamlit-mic-recorder